Made it so we use the JavaPlugin's logger rather then creating our own
got rid of the logger that we create by ourselves and made it so we use
the JavaPlugin's logger that we get when we inherit JavaPlugin. As a
result we can't make the log method static anymore, but we have access
to the JavaPlugin instance its fine. Also made it so that Database takes
a Logger param in the constructor and uses that rather then the static
log() method in JukeAlert that isn't there anymore