JukeAlert - Issue#6 - Fix for DB Bootstrapping Failure
Resolved a few issues that were preventing the database upgrade /
database installation for users.  Fixed issue with snitches table
reference not accepting the proper prefixed snitch table.  Fixed issue
with procedure attempting to be created prior to column existing.  Added
said column to the list of things not to throw an error for since that
is now the first item that would be found in said 'try catch' block.