Updated DetectorRailActivateListener
The file has been updated and improved and flattened. While some of the comments might seem a little obvious, that you can reasonably easily deduce what's going on, it's in my opinion better to save future developers, especially newer developers the effort of having to parse code. The comments allow me to define what's happening but also *why* it's happening.