NBTCache addition
New class that simplifies direct encoding to and decoding from nbt, acting as an interface for more complex data. This differs in how NBTCompound is an interface as that provides the means of setting a string to a key for example, whereas the extended class is meant to provide getters and setters along the lines of setMaterial() for example. The setter will encode the correct data directly onto the NBT instead of storing it in a field to be serialised later.