Revert "Alerting the player is a separate function now"
This reverts commit de91bb06ccff22ab1c083e470b86d97d8942fbc7.

This commit completely removes the use of aggregator alertUser
StringBuilder, and as a result of the constructs here will only alert
people of transforms / discoveries that have a customPrefix. If no
custom prefix for a drop is defined, no notification will ever be sent.
Aggregation of notifications is straight broken by this commit as well.

It also eliminates the " nearby" custom messaging of generation vs.
drop, just straight up gone?

I'll follow up this revert with a new commit to do the same but safely.