Fix another CivLogger bug, hopefully last one
It's weird because the issue I was having before was fix with the last change.. maybe PaperMC changed something? Either way, the nuclear option should the plugin be loaded but not instantiated is to create a replicant logger since PaperMC doesn't define the logger until after the plugin has been constructed.