Fix postgres healthcheck causing errors in log
pg_isready will use the current user to query the database, which
is the root user. This user does not exist in the database, which
causes "FATAL:  role "root" does not exist" to be logged every
time the healthcheck is run.

This commit fixes this issue with the -U flag to specify the user.
Note the $$ENV synax, see here for more info:
https://docs.docker.com/compose/environment-variables/