Add more description of civ at the beginning.
"A server based on nation building and political roleplay, where admin control is minimal and players have the opportunity to dispense their own justice. A world ripe with chance for any new friend to carve out their own legacy, in an almost decade long history. Filled with moments of high excitement in war and of peace, farming and socializing with your closest allies. Where countries, towns, and cities consist of more than just the blocks they're made of."