A partial fix for minecart derailment
If a minecart is going to derail because it is going to fast, this fix will backtrack its movement and move it as if it was going 8m/s. This will slow down the minecart slightly but prevent derailment.

This does not fix fast minecarts being unable going up hills or minecarts jumping across tracks.