STOP DOING HAS-THEN-GET!
This updates how ItemExchange does transactions. Instead of trying to determine whether the inventory has enough items, and then if-true, attempting the transaction which means iterating through the inventory again... instead this just immediately attempts the transaction (with cloned inventories, of course) and if it fails, then the transaction fails. This makes the control-flow of ItemExchange a lot simpler. If there are still bugs after this then I'm convinced that a Trisolaran sophon is messing with us.